+++ ChangeLog.md
@@ -11,6 +11,7 @@ This project's release branch is `master`. This log is written from the perspect
 * Add the `ob shell` command to enter a nix shell for an obelisk project
 * Allow skeleton's obelisk to be overridden. This changes the skeleton's default.nix interface: the arguments that it used to take are now part of the new "obelisk" argument.
 * Add `isoViaDecoded` and `isoViaEncoded` for extending isomorphisms to encoders.
+* Add `pairEncoders` to allow implementing an encoder with two simpler encoders and isomorphisms.
 
 ## v0.1.0.0 - 2019-03-29